Start to refactor writeside
