+++ components/core/src/streaming_archive/writer/GLTFile.cpp
