+++ components/core/src/streaming_archive/writer/GLTSegment.cpp
@@ -1,5 +1,5 @@
 #include "GLTSegment.hpp"
-#include "../../LogtypeSizeTracker.hpp"
+#include "../LogtypeSizeTracker.hpp"
 #include <iostream>
 
 using streaming_archive::glt::LogtypeSizeTracker;
