+++ components/core/src/streaming_archive/writer/LogtypeTable.cpp
